no need to specify the forward_net everytime in command line. the command line only overwrite the option in config file.
